Renamed "treetools" to just "trees"
