--- polymerist/tests/genutils/treetools/__init__.py
@@ -1,1 +1,0 @@
-'''Unit tests for `treetools` package'''
